Fix an unsigned->NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@31352 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSTextStorage.m
@@ -623,7 +623,7 @@ _attributesAtIndexEffectiveRange(
   return _textProxy;
 }
 
-- (NSDictionary*) attributesAtIndex: (unsigned)index
+- (NSDictionary*) attributesAtIndex: (NSUInteger)index
 		     effectiveRange: (NSRange*)aRange
 {
   unsigned dummy;
@@ -930,7 +930,7 @@ changeInLength: [aString length] - range.length];
 }
 
 // The superclass implementation is correct but too slow
-- (unsigned int) length
+- (NSUInteger) length
 {
   return [_textChars length];
 }
